+++ src/VernaForm.spec.tsx
@@ -4,33 +4,39 @@ import { JSONSchema7 } from 'json-schema';
 import _ from 'lodash';
 import { VernaContextProvider } from './VernaContextProvider';
 
-const schemaDefault: JSONSchema7 = {
-  title: 'A registration form',
-  description: 'Desc registration form',
-  type: 'object',
-  properties: {
-    testSection: {
-      type: 'object',
-      title: 'Sectiontest',
-      properties: {
-        champ1: {
-          type: 'string',
-          title: 'Field name 1',
-          propertyNames: true,
+describe('VernaForm', () => {
+  const getSchemaDefault = (): JSONSchema7 => ({
+    title: 'A registration form',
+    description: 'Desc registration form',
+    type: 'object',
+    properties: {
+      testSection: {
+        type: 'object',
+        title: 'Sectiontest',
+        properties: {
+          champ1: {
+            type: 'string',
+            title: 'Field name 1',
+          },
         },
       },
     },
-  },
-};
+  });
 
-describe('VernaForm', () => {
-  it('should render', async () => {
+  async function clickOnLastAddInputButton() {
+    _.last(
+      (await screen.findAllByRole('button', {
+        name: 'Add an input',
+      })) as HTMLElement[],
+    )?.click();
+  }
+
+  it('Should render', async () => {
     render(
-      <VernaContextProvider defaultSchema={schemaDefault}>
+      <VernaContextProvider defaultSchema={getSchemaDefault()}>
         <VernaForm />
       </VernaContextProvider>,
     );
-
     // - A fieldset legend should be displayed with form title
     screen.getByRole('group', { name: 'A registration form' });
 
@@ -44,6 +50,14 @@ describe('VernaForm', () => {
 
     expect($field1).toBeInstanceOf(HTMLInputElement);
     expect($field1.type).toBe('text');
+  });
+
+  it('Should be able to add or remove sections and fields', async () => {
+    render(
+      <VernaContextProvider defaultSchema={getSchemaDefault()}>
+        <VernaForm />
+      </VernaContextProvider>,
+    );
 
     // Add two sections
     screen
@@ -63,16 +77,8 @@ describe('VernaForm', () => {
     expect($addFieldButtons1).toHaveLength(3);
 
     // Add two input fields
-    _.last(
-      (await screen.findAllByRole('button', {
-        name: 'Add an input',
-      })) as HTMLElement[],
-    )?.click();
-    _.last(
-      (await screen.findAllByRole('button', {
-        name: 'Add an input',
-      })) as HTMLElement[],
-    )?.click();
+    await clickOnLastAddInputButton();
+    await clickOnLastAddInputButton();
 
     const $addFieldButtons2 = await screen.findAllByRole('button', {
       name: 'Add an input',
@@ -83,9 +89,9 @@ describe('VernaForm', () => {
       name: 'Add a section',
     });
     expect($addSectionButtons).toHaveLength(3);
+    screen.getByRole('button', { name: 'Submit' });
 
     // Delete every elements from top to bottom
-    screen.getByRole('button', { name: 'Submit' });
     _.forEach(
       (await screen.findAllByRole('button', {
         name: 'x',
@@ -104,5 +110,59 @@ describe('VernaForm', () => {
       name: 'Add an input',
     });
     expect($addFieldButtons3).toHaveLength(0);
+  });
+
+  it('Should use selector parameter to query sub schema and render it', async () => {
+    render(
+      <VernaContextProvider defaultSchema={getSchemaDefault()} defaultSelector={'testSection'}>
+        <VernaForm />
+      </VernaContextProvider>,
+    );
+
+    // - A fieldset legend should not be displayed
+    expect(screen.queryAllByRole('group', { name: 'A registration form' })).toHaveLength(0);
+
+    // - The form description should not be displayed
+    expect(screen.queryAllByText('Desc registration form')).toHaveLength(0);
+
+    // - The first section should be displayed
+    expect(screen.queryAllByRole('group', { name: 'Sectiontest' })).toHaveLength(1);
+
+    // - A required text input First name should be displayed inside the only section
+    const $field1 = document.getElementById('root_champ1') as HTMLInputElement;
+    expect($field1).toBeInstanceOf(HTMLInputElement);
+    expect($field1.type).toBe('text');
+  });
+
+  it('Should use selector parameter to query sub schema and add or remove fields on it', async () => {
+    render(
+      <VernaContextProvider defaultSchema={getSchemaDefault()} defaultSelector={'testSection'}>
+        <VernaForm />
+      </VernaContextProvider>,
+    );
+
+    // Add two input fields
+    await clickOnLastAddInputButton();
+    await clickOnLastAddInputButton();
+
+    const $addFieldButtons = await screen.findAllByRole('button', {
+      name: 'Add an input',
+    });
+    expect($addFieldButtons).toHaveLength(3);
+
+    // Delete every elements from top to bottom
+    _.forEach(
+      (await screen.findAllByRole('button', {
+        name: 'x',
+      })) as HTMLElement[],
+      (element) => element.click(),
+    );
+    const $deleteButtons = screen.queryAllByRole('button', { name: 'x' });
+    expect($deleteButtons).toHaveLength(0);
+
+    const $addFieldButtons3 = screen.queryAllByRole('button', {
+      name: 'Add an input',
+    });
+    expect($addFieldButtons3).toHaveLength(1);
   });
 });