+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Functionality to select and display only a section of the schema
 - Functionality to add or remove widgets and sections